% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UseMyTemplate.R
\name{UseMyTemplate}
\alias{UseMyTemplate}
\title{UseMyTemplate}
\usage{
UseMyTemplate(name = NULL, open = interactive(), ...)
}
\arguments{
\item{name}{name of analysis file}

\item{open}{should the file be opened after being created}

\item{...}{arguments to be passed to \link[usethis]{use_template}}
}
\description{
Use the MyTemplate Rmd for starting analysis
}
\examples{
\donttest{
if(interactive()){
  UseMyTemplate(name = "UseMyTemplate.Rmd")
}
}

}
